Change name of url-equal? to equivalent?
In my mind, two URIs being equal is a stronger relationship than mere
isomorphism.  Two different (not equal) URIs could possibly be
congruent, which is what this function tests for.

For example, the following two URIs are not the same, but should be
considered isomorphic:

  http://www.example.com?
  http://www.example.com/?

The name 'equivalent?' captures this fact better I think.

Also, there is no need for the 'url-' prefix.  See the Clojure coding
conventions wiki:

<http://www.assembla.com/wiki/show/clojure/Clojure_Library_Coding_Standards>